refactor: persistence
