+++ src/infra/entities/movie.entity.ts
